% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tp_function.R
\name{simulate_tp}
\alias{simulate_tp}
\title{A Monte Carlo simulation function}
\usage{
simulate_tp(x)
}
\description{
This function runs Monte Carlo simulations over a series of three-point estimates
}
